Multiple instances of FileCache resulted in disappearing keys.
Because the exact same file was used to keep track of keys the were overwritten when there were multiple instances of the same class with the same prefix.
Also different instances with different prefix also caused this because the filename was not unique.

It's now possible to instantiate multiple instances of the FileCache class with the same or different prefix, because the need of the "index" file has been removed an replaced by using the same key twice but with a -ttl prefix containing the ttl of the file.