Equal prefix because of normalizing.
When using for example a key named abd() and a key named abd## the "normalize" method created an equal cache key resulting in caching keys being overwritten. To prevent this we use sha1, which resolves this problem. However the files names are now also unreadable (just an fyi, imho this is not an issue).